+++ src/main/java/interfaces/BudgetType.java
@@ -1,11 +1,12 @@
 package interfaces;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public interface BudgetType extends Entity {
     BigInteger getBudgetTypeId();
-    BigInteger getGroupId();
+    Integer getGroupId();
     String getName();
     Boolean getRequired();
-    Double getCheckMax();
+    BigDecimal getCheckMax();
 }
